pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# -------------------------------------
# Poetry: Metadata
# -------------------------------------

[tool.poetry]
name = "aap-eda"
version = "0.1.0"
description = ""
authors = ["Red Hat, Inc. <info@ansible.com>"]
readme = "README.md"
packages = [{ include = "aap_eda", from = "src" }]

[tool.poetry.scripts]
aap-eda-manage = 'aap_eda.manage:main'

# -------------------------------------
# Poetry: Dependencies
# -------------------------------------
[tool.poetry.extras]
all = ["psycopg2"]
dev = ["psycopg2-binary"]

[tool.poetry.dependencies]
python = ">=3.9,<4"
django = "<4"
djangorestframework = "*"
dynaconf = ">=3"
drf-spectacular = "*"
channels = { version = "*", extras = ["daphne"] }
psycopg2 = { version = "*", optional = true }
psycopg2-binary = { version = "*", optional = true }
django-rq = "*"
rq = "*"
django-filter = "*"
pydantic = "*"
cryptography = "*"
kubernetes = "*"
podman = ">=4.5"
rq-scheduler = "^0.10"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-django = "*"
pytest-asyncio = "*"
requests = { version = "*", python = "<4.0" }
pytest-cov = "^4.1.0"
pytest-redis = "^3.0.2"

[tool.poetry.group.lint.dependencies]
flake8 = "*"
isort = "*"
black = "*"
flake8-broken-line = { version = "*", python = "<4.0" }
flake8-string-format = "*"
# This is an experimental linter.
ruff = "*"
# The rull claims that the flake8 plugins listed below are re-implemented,
#   These plugins will remain included until it's verified.
pep8-naming = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-docstrings = "*"
flake8-eradicate = { version = "*", python = "<4.0" }
flake8-print = "*"

# -------------------------------------
# Tools
# -------------------------------------

[tool.black]
line-length = 120
target-version = ["py39", "py310"]

[tool.isort]
profile = "black"
combine_as_imports = true
line_length = 120

[tool.ruff]
line-length = 120
select = [
    "E",
    "F",
    "D",   # flake8-docstrings
    "TID", # flake8-tidy-imports
]
extend-ignore = [
    "D1", #  Missing docstrings errors
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"src/aap_eda/core/migrations/*" = ["E501"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.pydocstyle]
convention = "pep257"
